Use double samples in AddMeasurement
AddMeasurement method from Result class adds a new
entry in the samples map (the key being the name of
the measurement while the value is the measurement
samples vector).

Samples data type is changed from long int to
double in order to provide a better accuracy.
Keeping samples as doubles improves the precision
when computing the measurements statistics.

Bug: 195113227
Test: manual

+++ src/instruction_set.cpp
@@ -105,7 +105,7 @@ void InstructionSet::RunInstructions() {
 // The duration for an instruction set is a sum of its child instruction durations.
 std::unique_ptr<Result> InstructionSet::CollectResults(const std::string& prefix) {
   auto result = std::make_unique<Result>(prefix + name_, repeat_);
-  std::vector<int64_t> duration;
+  std::vector<double> duration;
   for (const auto& instruction : instructions_) {
     auto sub_result = instruction->CollectResults("");
     auto samples = sub_result->GetSamples("duration");
